Spike of a load generator and simple monitoring UI.
This is to make it easy for users to run simple load tests and see how things
perform. Ideally this will be complemented by a prometheus (or similar) setup
to have full access to our metrics.

However, I've added a simple stats endpoint to the main application API in
https://github.com/temporalio/reference-app-orders-go/pull/87 so that we can at
least show some monitoring in the UI itself without needing the user to deploy
and configure the metrics system.